signature registered party
Update mithril-stm/src/single_signature/mod.rs
