+++ mithril-stm/src/lib.rs
@@ -16,10 +16,9 @@ pub use error::{
 };
 pub use key_reg::{ClosedKeyReg, KeyReg};
 pub use participant::{StmInitializer, StmSigner, StmVerificationKey, StmVerificationKeyPoP};
-pub use single_signature::StmSig;
+pub use single_signature::{StmSig, StmSigRegParty};
 pub use stm::{
     CoreVerifier, Index, Stake, StmAggrSig, StmAggrVerificationKey, StmClerk, StmParameters,
-    StmSigRegParty,
 };
 
 #[cfg(feature = "benchmark-internals")]
